RTC Fully Integrated
Update Parent Device main code with RTC functionality fully integrated. The RTC is able to trigger an interrupt using an alarm on the I/O (SQW) pin. The microcontroller is then able to get the current time stored in the RTC registers and clear the alarm flag so it can occur again on the next interval.